backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cartrack-backend"
version = "0.1.0"
description = "CAR TRACK backend (FastAPI)"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "pydantic>=2.7",
  "python-dotenv>=1.0.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
  "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["app"]
